- added Hibernate configuration with sample test user

+++ src/main/java/com/voidmainworld/plutus/vo/user/UserVO.java
@@ -9,20 +9,20 @@ import org.codehaus.jackson.map.annotate.JsonSerialize;
 @JsonSerialize(include = JsonSerialize.Inclusion.NON_NULL)
 public class UserVO implements Serializable {
 
-	private static final long serialVersionUID = 7456046476491395764L;
+	private static final long	serialVersionUID	= 7456046476491395764L;
 
-	private long id;
-	private String name;
+	private int					id;
+	private String				name;
 
 	public UserVO() {
 		super();
 	}
 
-	public UserVO(long id) {
+	public UserVO(int id) {
 		this(id, Long.toString(id));
 	}
 
-	public UserVO(long id, String name) {
+	public UserVO(int id, String name) {
 		super();
 		this.id = id;
 		this.name = name;
@@ -32,7 +32,7 @@ public class UserVO implements Serializable {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
